PWA-1906 Moved interface to API, adding missing docblocks, refactor for maintainability
