+++ Model/Computed/PageInfo.php
@@ -1,10 +1,15 @@
 <?php
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
 
 declare(strict_types=1);
 
 namespace Magento\UpwardConnector\Model\Computed;
 
 use Magento\Upward\DefinitionIterator;
+use Magento\UpwardConnector\Api\ComputedInterface;
 use Magento\UpwardConnector\Model\PageType;
 
 class PageInfo implements ComputedInterface
@@ -18,6 +23,11 @@ class PageInfo implements ComputedInterface
         $this->pageTypeResolver = $pageTypeResolver;
     }
 
+    /**
+     * @param \Magento\Upward\DefinitionIterator $iterator
+     *
+     * @return string
+     */
     public function resolve(DefinitionIterator $iterator)
     {
         return $this->pageTypeResolver
