+++ Model/ComputedPool.php
@@ -1,14 +1,18 @@
 <?php
+/**
+ * Copyright © Magento, Inc. All rights reserved.
+ * See COPYING.txt for license details.
+ */
 
 declare(strict_types=1);
 
 namespace Magento\UpwardConnector\Model;
 
-use Magento\UpwardConnector\Model\Computed\ComputedInterface;
+use Magento\UpwardConnector\Api\ComputedInterface;
 
 class ComputedPool
 {
-    /** @var \Magento\UpwardConnector\Model\Computed\ComputedInterface[] */
+    /** @var \Magento\UpwardConnector\Api\ComputedInterface[] */
     private $items;
 
     public function __construct(?array $items = [])
@@ -16,6 +20,12 @@ class ComputedPool
         $this->items = $items;
     }
 
+    /**
+     * Get resolving ComputedInterface class
+     *
+     * @param string $classKey
+     * @return \Magento\UpwardConnector\Api\ComputedInterface|null
+     */
     public function getItem(string $classKey): ?ComputedInterface
     {
         return $this->items[$classKey] ?? null;
